requirements.txt
mediapipe>=0.10.32
opencv-python>=4.10.0
numpy>=2.0.0
flet>=0.24.0
websockets>=16.0
pystray>=0.19.5
Pillow>=10.0.0
pyperclip>=1.9.0
